Create mux_select_line_circuit
It generates the mux output according to the mux selection line. mux selection circuit uses the input from the fault_tolerant_ripple_carry_adder and compares with the stored value of the look up table or expected output of the ripple carry adder.If there are any mismatch between the expected output and actual output then the mux select line bypasses the selected full adder.